Exception handling - Delete

+++ src/main/java/com/julio/SBProject/services/UserService.java
@@ -5,10 +5,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.julio.SBProject.entites.User;
 import com.julio.SBProject.repositories.UserRepository;
+import com.julio.SBProject.services.exceptions.DatabaseException;
 import com.julio.SBProject.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,7 +33,14 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		try {
+			if (!userRepository.existsById(id)) throw new ResourceNotFoundException(id);			
+			userRepository.deleteById(id);			
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 	
 	public User update(Long id, User obj) {
@@ -39,7 +49,7 @@ public class UserService {
 		return userRepository.save(entity);
 	}
 
-	private void updateData(User entity, User obj) {
+	private void updateData(User entity, User obj) { 
 		entity.setName(obj.getName());
 		entity.setEmail(obj.getEmail());
 		entity.setPhone(obj.getPhone());
